+++ examples/tcp/server.rs
@@ -20,8 +20,8 @@ pub fn run() {
 
     let listen_addr = "127.0.0.1:3000";
     match network.listen_tcp(listen_addr) {
-        Some(_) => println!("TCP Server running at {}", listen_addr),
-        None => return println!("Can not listening at selected interface/port"),
+        Ok(_) => println!("TCP Server running at {}", listen_addr),
+        Err(_) => return println!("Can not listening at selected interface/port"),
     }
 
     loop {
@@ -30,7 +30,7 @@ pub fn run() {
                 NetEvent::Message(client_id, message) => match message {
                     Message::Greetings(text) => {
                         println!("Client ({}) says: {}", clients[&client_id], text);
-                        network.send(client_id, Message::Greetings("Hi, I hear you".into()));
+                        network.send(client_id, Message::Greetings("Hi, I hear you".into())).unwrap();
                     },
                 },
                 NetEvent::AddedEndpoint(client_id, addr) => {